play-web/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.11.5
aiofiles==23.2.1
python-magic==0.4.27
